refactor(ui/layout): move code files to a folder
BREAKING CHANGE: add a new path
